requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv